+++ src/http/middleware/extract_external_token/external_token_error.rs
@@ -7,13 +7,9 @@ use actix_web::dev::ServiceRequest;
 /// missing `Authorization` header and token parsing/validation failure.
 pub trait ExternalTokenError: ResponseError {
     /// Builds an error for requests that do not include an `Authorization` header.
-    fn external_token_not_present(request: &ServiceRequest) -> Self;
+    fn token_not_present(request: &ServiceRequest, internal: bool) -> Self;
 
     /// Builds an error for requests where the `Authorization` header exists
     /// but the token cannot be extracted or parsed.
-    fn token_extraction_failed(request: &ServiceRequest, cause: anyhow::Error) -> Self;
-
-    /// Builds an error for the cases when the HTTP pipeline encounters a duplicated
-    /// [`EncryptedToken`] in the request.
-    fn token_not_present(request: &ServiceRequest) -> Self;
+    fn token_extraction_failed(request: &ServiceRequest, internal: bool, cause: anyhow::Error) -> Self;
 }
